chiti_organized/backend/requirements.txt
# Team Chitti RAG - Requirements
# Modular RAG Q&A System with Online/Local/Hybrid modes
# Updated for January 2026

# Core dependencies
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1

# Online Mode Dependencies
# Azure OpenAI + OpenAI
openai>=1.35.0

# Qdrant Vector Database
qdrant-client>=1.9.0

# MongoDB Atlas
pymongo>=4.8.0
motor>=3.5.0  # Async MongoDB driver

# Local Mode Dependencies
# Phi-3 Mini LLM
torch>=2.3.0
transformers>=4.42.0
accelerate>=0.32.0

# Local Embeddings
sentence-transformers>=3.0.0

# FAISS Vector Store
faiss-cpu>=1.8.0  # Use faiss-gpu if you have CUDA

# Utilities
numpy>=2.0.0
scipy>=1.14.0
scikit-learn>=1.5.0

# Text Processing
nltk>=3.9.0
spacy>=3.8.0
textstat>=0.7.0

# Document Processing
pypdf>=4.3.0
python-docx>=1.2.0
markdown>=3.7.0
beautifulsoup4>=4.13.0
python-multipart>=0.0.9

# Web Framework
jinja2>=3.1.0
aiofiles>=24.1.0

# Logging and Monitoring
loguru>=0.8.0
prometheus-client>=0.21.0

# Development and Testing
pytest>=8.3.0
pytest-asyncio>=0.23.0
black>=24.0.0
isort>=5.13.0
flake8>=7.1.0

# Optional: GPU acceleration (uncomment if needed)
# torch>=2.3.0+cu121 --index-url https://download.pytorch.org/whl/cu121
# faiss-gpu>=1.8.0